Separate components into separate files.
